+++ src/App.tsx
@@ -1,5 +1,5 @@
 import './App.css'
-import Header from './header'
+import Header from './components/header'
 import Content from './components/Content'
 import Navbar from './Navbar'
