+++ src/components/header.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import './header.css';
-import image from './assets/gossipgirl.png';
+import '../styles/header.css';
+import image from '../assets/gossipgirl.png';
 
 const Header: React.FC = () => {
     return (
